Change the way special characters are handled
 - disable word expansion / insertion of surplus chars

Previously, single characters we wanted to ignore were represented with carets (^) in the plaintext, and the handler replaced each of them with an ignored single quote. This scheme is limited in multiple ways:
- all carets will be transformed into ignored single quotes, even if they're meant to be displayed as regular carets
- only single quotes can be marked as ignored this way, because carets and ignored single quotes are in a 1-to-1 mapping

The new scheme:
dollar sign ($): escapes next char in the paragraph, i.e. next char is handled as a regular letter with no special behavior attached
caret (^): ignore next character in paragraph
This fixes the limitations of the previous scheme: carets can be dispayed by prefixing them with a '$', and any character can be ignore by prefixing it with a '^'. This method also has some limitations (e.g. the syntax for ignoring a whole word or paragraph is unpleasant), but for our current use case it does a good enough job.

(there's probably an npm package I could have used for this)