Implement 'word appension/shrinking' mechanism
Whenever the user inputs an incorrect character where a space
is expected, the app will append the character to the word behind.
The character becomes a 'surplus' character.

In relation to this, when the user inputs a backspace in front of a
surplus character, the character will be removed from the text.

Texts containing multiple consecutive spaces were not tested and
are not supported.