Implement API authentication
Authentication requests are routed to AuthController. Registered
users' identity information is stored in persistent storage.

A JSON web token is returned in the response body on a successful login.
The token will act as a bearer token on further requests by the client,
and will grant access to a wider range of the API. The token signature
contains the server's secret symmetric key and is signed using the
HMAC-SHA256 algorithm.